added level and movement
